requirements.txt
google-generativeai
gradio
pandas
python-docx
openpyxl
spacy
requests
dotenv